ClickHouseIO - Improve user agent (#37815)
* product name is deprecated, changing to client name

* create a share method + append external client name

* add unittests and rename original test file to IT

* reuse the same properties

* add user agent logic to the generated documentation

* fix query log targeted column

* add integration test to verify user agent is correct

* make sure integration tests are included

* move IT out of test task
