+++ sdks/java/io/clickhouse/src/main/java/org/apache/beam/sdk/io/clickhouse/ClickHouseIO.java
@@ -107,6 +107,18 @@ import org.slf4j.LoggerFactory;
  * href="https://clickhouse.com/docs/guides/developer/deduplication">Deduplication strategies
  * documentation</a>
  *
+ * <h4>User agent</h4>
+ *
+ * <p>The connector automatically sets the ClickHouse client name to {@code Apache Beam/<version>},
+ * which is visible in {@code system.query_log.client_name}. If you set the {@code client_name}
+ * connection property, it is appended after the Beam identifier, for example:
+ *
+ * <pre>{@code
+ * Properties props = new Properties();
+ * props.setProperty("client_name", "MyApp/1.0");
+ * // Results in: "Apache Beam/<version> MyApp/1.0"
+ * }</pre>
+ *
  * <h4>Mapping between Beam and ClickHouse types</h4>
  *
  * <table summary="Type mapping">
@@ -244,21 +256,19 @@ public class ClickHouseIO {
 
     @Override
     public PDone expand(PCollection<T> input) {
-      TableSchema tableSchema = tableSchema();
-      if (tableSchema == null) {
-        tableSchema = getTableSchema(clickHouseUrl(), database(), table(), properties());
-      }
-
-      String sdkVersion = ReleaseInfo.getReleaseInfo().getSdkVersion();
-      String userAgent = String.format("Apache Beam/%s", sdkVersion);
 
       Properties properties = properties();
+      set(properties, "client_name", buildClientName(properties));
+
+      TableSchema tableSchema = tableSchema();
+      if (tableSchema == null) {
+        tableSchema = getTableSchema(clickHouseUrl(), database(), table(), properties);
+      }
 
       set(properties, "max_insert_block_size", maxInsertBlockSize());
       set(properties, "insert_quorum", insertQuorum());
       set(properties, "insert_distributed_sync", insertDistributedSync());
       set(properties, "insert_deduplication", insertDeduplicate());
-      set(properties, "product_name", userAgent);
 
       WriteFn<T> fn =
           new AutoValue_ClickHouseIO_WriteFn.Builder<T>()
@@ -526,8 +536,7 @@ public class ClickHouseIO {
               .setPassword(password)
               .setDefaultDatabase(database())
               .setOptions(options)
-              .setClientName(
-                  String.format("Apache Beam/%s", ReleaseInfo.getReleaseInfo().getSdkVersion()));
+              .setClientName(properties().getProperty("client_name"));
 
       // Add optional compression if specified in properties
       String compress = properties().getProperty("compress", "false");
@@ -725,8 +734,7 @@ public class ClickHouseIO {
               .setUsername(user)
               .setPassword(password)
               .setDefaultDatabase(database)
-              .setClientName(
-                  String.format("Apache Beam/%s", ReleaseInfo.getReleaseInfo().getSdkVersion()));
+              .setClientName(buildClientName(properties));
 
       try (Client client = clientBuilder.build()) {
         String query = "DESCRIBE TABLE " + quoteIdentifier(table);
@@ -764,6 +772,17 @@ public class ClickHouseIO {
     } catch (Exception e) {
       throw new RuntimeException("Failed to get table schema for table: " + table, e);
     }
+  }
+
+  @VisibleForTesting
+  static String buildClientName(Properties properties) {
+    String beamAgent =
+        String.format("Apache Beam/%s", ReleaseInfo.getReleaseInfo().getSdkVersion());
+    String existingClientName = properties.getProperty("client_name");
+    if (!Strings.isNullOrEmpty(existingClientName)) {
+      return beamAgent + " " + existingClientName;
+    }
+    return beamAgent;
   }
 
   static String quoteIdentifier(String identifier) {
